requirements.txt
matplotlib==3.9.1
numpy==2.0.0
Pillow==10.4.0
tensorflow==2.17.0
